server/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
paho-mqtt==2.0.0
scikit-learn==1.4.1.post1
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
